24.09.04 (수) 더 마르스
오늘 한 일
카드 획득 아이콘
클릭시 카드 선택 UI 활성화
카드 획득 UI
랜덤으로 나온 3장의 카드 중 선택해서 얻는 기능
배속 적용 (1, 2, 3)배속

+++ TheMars/Assets/OnwStudio/Scripts/MainGame/Building/Effect/Creator/GetMarsLithiumEffectCreator.cs
@@ -10,13 +10,13 @@ namespace TM.Building.Effect.Creator
     public class GetMarsLithiumEffectCreator : ITMBuildingEffectCreator
     {
         // .. TODO : 구조 수정
-        [field: SerializeField, Min(1), DisplayAs("Level 1 반복 시간")] public int LevelOneRepeatSeconds { get; private set; }
+        [field: SerializeField, Min(1), DisplayAs("Level 1 반복 시간")] public float LevelOneRepeatSeconds { get; private set; }
         [field: SerializeField, Min(1), DisplayAs(("Level 1 마르스 리튬"))] public int LevelOneMarsLithium { get; private set; }
         
-        [field: SerializeField, Min(1), DisplayAs("Level 2 반복 시간")] public int LevelTwoRepeatSeconds { get; private set; }
+        [field: SerializeField, Min(1), DisplayAs("Level 2 반복 시간")] public float LevelTwoRepeatSeconds { get; private set; }
         [field: SerializeField, Min(1), DisplayAs(("Level 2 마르스 리튬"))] public int LevelTwoMarsLithium { get; private set; }     
         
-        [field: SerializeField, Min(1), DisplayAs("Level 3 반복 시간")] public int LevelThreeRepeatSeconds { get; private set; }
+        [field: SerializeField, Min(1), DisplayAs("Level 3 반복 시간")] public float LevelThreeRepeatSeconds { get; private set; }
         [field: SerializeField, Min(1), DisplayAs(("Level 3 마르스 리튬"))] public int LevelThreeMarsLithium { get; private set; }
         
         public ITMBuildingEffect CreateEffect()
